requirements.txt
twelvelabs
python-dotenv
boto3==1.35.95
celery==5.4.0
redis==5.2.1
requests==2.32.3
beautifulsoup4==4.12.3
Django==5.1.5
djangorestframework==3.15.2
django-cors-headers==4.6.0
psycopg[binary]==3.2.13
